+++ CHANGELOG.md
@@ -1,2 +1,6 @@
 ## 0.1.0
 * Initial release
+
+## 0.1.1
+* Fix schema warning message format
+* Make schema warning display optional, hidden by default